# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "本地化的 Bangumi 数据服务。利用 Archive 每周导出的 wiki 数据导入本地 SQLite，\n提供条目/人物/角色/章节等 REST 查询接口与单二进制管理接口。\n统一响应包装：成功 {\"ok\": true, \"data\": ...}；失败 {\"ok\": false, \"error\": \"...\"}。\n数据接口均为只读；管理接口（/api/admin/*）需要 X-Admin-Token 请求头（未配置 token 时放行）。\n交互式文档：/docs 或 /swagger/index.html；本文档由 swag 从代码注解自动生成（make docs）。",
        "title": "Bangumi 本地数据搜索 API",
        "contact": {
            "name": "GitHub",
            "url": "https://github.com/nz3u/bangumi-data"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/episodes/search": {
            "get": {
                "description": "q 同时命中章节标题与所属条目标题（原名/中文名/别名，归一化口径同条目搜索）；字段含义同 Archive 的 episode 表：sort=集数、disc=所在光盘、airdate=播出时间（原文文本，范围匹配对 ISO 日期有效）、duration=时长。≥3 字符走 trigram 全文索引，短词自动走命中集路径。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "章节"
                ],
                "summary": "章节搜索与筛选",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键词（全文搜索）",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "所属条目 ID",
                        "name": "subject_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "作品类型（所属条目）：1 书籍 / 2 动画 / 3 音乐 / 4 游戏 / 6 三次元",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2,
                            3,
                            4,
                            5,
                            6
                        ],
                        "type": "integer",
                        "description": "章节类型：0 正篇 / 1 SP / 2 OP / 3 ED / 4 Trailer / 5 MAD / 6 其他",
                        "name": "ep_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "光盘号",
                        "name": "disc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "播出时间下界",
                        "name": "airdate_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "播出时间上界",
                        "name": "airdate_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序：id / sort=集数 / airdate=播出日期 / subject=条目 / popularity=条目人气，省略=智能排序",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "排序方向",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.episodeSearchData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.apiEnvelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
...
        "api.episodeBrief": {
            "type": "object",
            "properties": {
                "airdate": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "disc": {
                    "type": "integer"
                },
                "duration": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "name_cn": {
                    "type": "string"
                },
                "sort": {
                    "type": "number"
                },
                "subject": {
                    "$ref": "#/definitions/api.episodeSubjRef"
                },
                "type": {
                    "type": "integer"
                },
                "type_name": {
                    "type": "string"
                }
            }
        },
        "api.episodeSearchData": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.episodeBrief"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.episodeSubjRef": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "name_cn": {
                    "type": "string"
                },
                "platform_name": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "type_name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminToken": {
            "type": "apiKey",
            "name": "X-Admin-Token",
            "in": "header"
        }
    }
}